Splited main file
