+++ Sources/consts_exemple.py
