+++ Sources/main.py
@@ -1,222 +1,24 @@
-# import urequests
-from urequests import request
-import network
-import framebuf
-from math import ceil
-import esp32
-import utime
-import ujson
-import _thread
+# Libs
 import machine
-import ssd1306
+import esp32
+from math import ceil
+import utime
+
+# Local libs
+import uasyncio as asyncio
+
+# Local scripts
+from screen import Screen_element
+from screen import Screen_Handler
+from internet import Network
 import consts as const
 
+# import _thread
+# import framebuf
 
-class Screen_Handler:
-    def __init__(self):
-        # Constants
-        self.screen_columns = 16
-        self.screen_spacing = 8
-        self.screen_width = 128
-        self.screen_height = 64
-        self.char_width = int(self.screen_width / self.screen_columns)
-        self.char_height = int(self.screen_height / self.screen_spacing)
-
-        # set reset Pin hight
-        # try:
-        # pin16 = machine.Pin(16, machine.Pin.OUT)
-        # pin16.value(1)
-        # machine.Pin(16, machine.Pin.OUT).value(1)
-        # self.i2c = machine.I2C(scl=machine.Pin(15), sda=machine.Pin(4))
-
-        self.i2c = machine.I2C(scl=machine.Pin(4), sda=machine.Pin(5))
-
-        # self.i2c = machine.I2C(scl=machine.Pin(15), sda=machine.Pin(4))
-        # print(self.i2c.scan())
-        self.oled = ssd1306.SSD1306_I2C(self.screen_width, self.screen_height, self.i2c)
-        self.oled.fill(0)
-        self.memory_index = {}
-
-        # Array of functions for displayable elements
-        self.displayables = {
-            "str": self.display_str,
-            "pixel": self.display_pixel,
-            "line": self.display_line,
-            "rect": self.display_rect,
-        }
-
-        # Pixel arts
-        self.pixel_art = {
-            "up_arrow": [
-                "001100",
-                "011110",
-                "111111",
-                "001100",
-                "001100",
-                "001100",
-                "001100",
-            ],
-            "cross": [
-                "100001",
-                "110011",
-                "011110",
-                "001100",
-                "011110",
-                "110011",
-                "100001",
-            ],
-            "check": [
-                "000001",
-                "000001",
-                "000011",
-                "000010",
-                "110110",
-                "011100",
-                "001100",
-            ],
-            "thunder": [
-                "011110",
-                "111111",
-                "111111",
-                "011110",
-                "000100",
-                "001100",
-                "001000",
-            ],
-            "rain": [
-                "011110",
-                "111111",
-                "111111",
-                "011110",
-                "010010",
-                "001001",
-                "010010",
-            ],
-            "snow": [
-                "011110",
-                "111111",
-                "111111",
-                "011110",
-                "101010",
-                "010101",
-                "101010",
-            ],
-            "mist": [
-                "011110",
-                "111111",
-                "111111",
-                "011110",
-                "000111",
-                "111000",
-                "000111",
-            ],
-            "clear": [
-                "011110",
-                "111111",
-                "111111",
-                "111111",
-                "111111",
-                "011110",
-                "000000",
-            ],
-            "clouds": [
-                "001110",
-                "011111",
-                "011111",
-                "110001",
-                "100001",
-                "100001",
-                "011110",
-            ],
-            "celcius": [
-                "011000",
-                "100000",
-                "100000",
-                "011000",
-                "000000",
-                "000000",
-                "000000",
-            ],
-        }
-
-    def width_to_pixel(self, x):
-        return int(self.screen_width / self.screen_columns * x)
-
-    def height_to_pixel(self, y):
-        return int(self.screen_height / self.screen_spacing * y)
-
-    def reset_zone(self, x1, y1, x2, y2):
-        self.oled.rect(x1, y1, x2 - x1, y2 - y1, True, 0)
-
-    def display_str(self, elem):
-        x1, y1, string = elem
-        x1 = self.width_to_pixel(x1)
-        y1 = self.height_to_pixel(y1)
-        x2 = len(string) * self.char_width
-        y2 = self.char_height
-        self.reset_zone(x1, y1, x2, y2)
-        self.oled.text(string, x1, y1)
-        return (x1, y1, x2, y2)
-
-    def display_pixel(self, elem):
-        x1, y1, content_name = elem
-        x1 = self.width_to_pixel(x1)
-        y1 = self.height_to_pixel(y1)
-        art = self.pixel_art[content_name]
-        self.reset_zone(x1, y1, len(art[0]), len(art))
-        y2 = y1
-        for pixel_str in self.pixel_art[content_name]:
-            x2 = x1
-            for pixel in pixel_str:
-                if pixel == "1":
-                    self.oled.pixel(x2, y2, 1)
-                x2 += 1
-            y2 += 1
-        return (x1, y1, x2, y2)
-
-    def display_line(self, elem):
-        x1, y1, x2, y2 = elem
-        self.reset_zone(x1, y1, x2, y2)
-        self.oled.line(x1, y1, x2, y2)
-        return (x1, y1, x2, y2)
-
-    def display_rect(self, elem):
-        x1, y1, x2, y2, fill, col = elem
-        self.reset_zone(x1, y1, x2, y2)
-        self.oled.rect(x1, y1, x2 - x1, y2 - y1, fill, col)
-        return (x1, y1, x2, y2)
-
-    def set_memory(self, name, elem_type=None, content=None, delete=False):
-        # Delete element
-        if delete and name in self.memory_index:
-            x1, y1, x2, y2 = self.memory_index[name]
-            self.reset_zone(x1, y1, x2, y2)
-            del self.memory_index[name]
-        if elem_type is not None and content is not None:
-            elems = tuple(list(content))
-            self.memory_index[name] = self.displayables[elem_type](content)
-
-    def update_display(self):
-        self.oled.show()
-
-
-class Screen_element:
-    def __init__(self, ntw, sc, max_time_check):
-        self.ntw = ntw
-        self.sc = sc
-        self.last_time_check = 0
-        self.time_diff = 0
-        self.max_time_check = max_time_check  # * 1000
-
-    def get_time_diff(self, now):
-        self.time_diff = now - self.last_time_check
-        if self.time_diff < 0:
-            self.last_time_check = 0
-            self.time_diff = now
-        if self.time_diff > self.max_time_check:
-            self.last_time_check = now
-            return True
-        return False
+# https://docs.python.org/3.5/library/_thread.html#module-_thread
+# https://github.com/loboris/MicroPython_ESP32_psRAM_LoBo/wiki/thread_example_1
+# _thread.start_new_thread(google.get, (now,))
 
 
 class Clock(Screen_element):
@@ -254,6 +56,24 @@ class Clock(Screen_element):
             name="date", elem_type="str", content=(1, 0, time + " " + date)
         )
 
+    async def toggle_get_async():
+        while True:
+            await asyncio.sleep_ms(const.MAIN_CYCLE_TIME)
+            if self.get_time_diff(utime.time()) or self.tm is None:
+                self.set()
+            localtime = utime.localtime()
+            date = " " + "%02d" % localtime[2] + "/" + "%02d" % localtime[1]
+            time = (
+                "%02d" % localtime[3]
+                + ":"
+                + "%02d" % localtime[4]
+                + ":"
+                + "%02d" % localtime[5]
+            )
+            self.sc.set_memory(
+                name="date", elem_type="str", content=(1, 0, time + " " + date)
+            )
+
 
 class Weather(Screen_element):
     def __init__(self, ntw, sc, max_time_check):
@@ -281,6 +101,7 @@ class Weather(Screen_element):
         if self.get_time_diff(now) or self.current_temperature is None:
             print("Getting weather")
             WEATHER_data = self.ntw.request("GET", self.complete_url)
+            print("Weather response =", WEATHER_data)
 
             if WEATHER_data is not None and WEATHER_data["cod"] != "404":
                 self.WEATHER_description = WEATHER_data["weather"][0]["description"]
@@ -322,89 +143,6 @@ class Weather(Screen_element):
                 elem_type="str",
                 content=(2 + len(temp_str), y, str(self.current_humidity) + "%"),
             )
-
-
-class Network(Screen_element):
-    def __init__(self, sc, max_time_check):
-        Screen_element.__init__(self, None, sc, max_time_check)
-        self.ip = None
-        self.wlan = network.WLAN(network.STA_IF)
-        self.wlan.active(True)
-        self.trying_to_connect = False
-        self.ssid = None
-        self.pswd = None
-
-    def request(self, request_type, url, data=None, headers={}):
-        if self.wlan.isconnected():
-            json = None
-            if data is not None:
-                data = ujson.dumps(data)
-            response = request(request_type, url, data=data, headers=headers)
-            if response is not None:
-                json = response.json()
-            response.close()
-            return json
-        print("request error")
-        return None
-
-    def connect(self):
-        self.sc.set_memory(
-            name="connection_status", elem_type="pixel", content=(0, 0, "cross")
-        )
-        self.sc.update_display()
-        print("connecting to:", self.ssid, " with password ", self.pswd)
-        self.wlan.connect(self.ssid, self.pswd)
-
-    def get_best_wifi(self):
-        available_wifi = []
-
-        # Scan for available networks and only keep known ones
-        for wifi in self.wlan.scan():
-            ssid = wifi[0].decode("utf-8")
-            signal_strenght = wifi[3]
-            if ssid in const.NTW_LIST:
-                available_wifi.append((ssid, signal_strenght))
-
-        # Only keep the one with the best signal
-        if len(available_wifi) > 0:
-            print("Available wifi ", available_wifi)
-            self.ssid = max(available_wifi, key=lambda wifi: wifi[1])[0]
-            self.pswd = const.NTW_LIST[self.ssid]
-            print("Select wifi ", self.ssid)
-        else:
-            print("No wifi available")
-            self.ssid = None
-            self.pswd = None
-
-        return self.ssid
-
-    def check_connection(self, now):
-        # Check if we have to check for connection
-        if (
-            self.get_time_diff(now)
-            or not self.wlan.isconnected()
-            or self.trying_to_connect
-        ):
-            # print("checking connection")
-            self.last_time_check = now
-
-            # If not connected try connecting
-            if not self.wlan.isconnected() and not self.trying_to_connect:
-                print("connecting")
-                if self.get_best_wifi() is not None:
-                    self.connect()
-                    self.trying_to_connect = True
-
-            # If we have reconnected last iteration then update
-            if self.wlan.isconnected():
-                self.ip = self.wlan.ifconfig()
-                print("network config:", self.ip)
-                self.sc.set_memory(
-                    name="connection_status", elem_type="pixel", content=(0, 0, "check")
-                )
-                self.trying_to_connect = False
-
-        return self.wlan.isconnected()
 
 
 class Google(Screen_element):
@@ -624,11 +362,15 @@ class Google(Screen_element):
                     )
 
 
+# async def killer(duration):
+#    await asyncio.sleep(duration)
+
+
 def main():
     sc = Screen_Handler()
     ntw = Network(sc, const.NTW_CHECK_TIME)
-    weather = Weather(ntw, sc, const.WEATHER_TIME_CHECK)
     clock = Clock(ntw, sc, const.CLOCK_TIME_CHECK)
+    weather = Weather(ntw, sc, const.WEATHER_TIME_CHECK)
     google = Google(ntw, sc)
     sc.set_memory(
         name="line_top",
@@ -655,35 +397,24 @@ def main():
         ),
     )
 
-    # https://docs.python.org/3.5/library/_thread.html#module-_thread
-    # https://github.com/loboris/MicroPython_ESP32_psRAM_LoBo/wiki/thread_example_1
-    # _thread.start_new_thread(ok, ("in thread",))\
-
     # https://github.com/peterhinch/micropython-async/blob/master/TUTORIAL.md#01-installing-uasyncio-on-bare-metal
-    import uasyncio as asyncio
+    # https://forum.micropython.org/viewtopic.php?f=2&t=2876&start=10
+    # https://github.com/peterhinch/micropython-mqtt/tree/master/mqtt_as
 
-    async def bar():
-        count = 0
-        while True:
-            count += 1
-            print(count)
-            await asyncio.sleep(1)  # Pause 1s
+    # loop = asyncio.get_event_loop()
+    # loop.create_task(clock.toggle_get_async())
+    # loop.run_until_complete(killer(10))
+    # loop.close()
+    # return
 
-    loop = asyncio.get_event_loop()
-    loop.create_task(bar())  # Schedule ASAP
-    loop.run_forever()
-    print("here")
-
+    print("Running main loop")
     while True:
         now = utime.time()
         is_connected = ntw.check_connection(now)
         if is_connected:
-            weather.get(now)
             clock.get(now)
-            google.get(now)
-            # _thread.start_new_thread(google.get, (now,))
-            # else:
-            #     print("No internet connection")
+            weather.get(now)
+            # google.get(now)
         sc.update_display()
         utime.sleep(const.MAIN_CYCLE_TIME)
